==> notebooks/1_data_cleaning.ipynb <==
# Settings

RAW_OUTPUT_FOLDER = "../crawler/ctsv_scraper/output"

# Basic Statistics, Remove Not Exist Events, Deleted Events 

import os 
os.listdir(RAW_OUTPUT_FOLDER) # 2 files

import json

# Read input data
act_data = []
with open(os.path.join(RAW_OUTPUT_FOLDER, "activities.jsonl")) as f:
    for line in f:
        act_data.append(json.loads(line))
    
usr_data = []
with open(os.path.join(RAW_OUTPUT_FOLDER, "user_activities.jsonl")) as f:
    for line in f:
        usr_data.append(json.loads(line))

# Activity data
len(act_data), act_data[0]

act_data.sort(key=lambda x: x["AId"])
act_data[-1]

# User data
len(act_data), act_data[0]

usr_data.sort(key=lambda x: x['AId'])
len(usr_data), usr_data[-1]

for sample in usr_data:
    if sample["AId"] == 8886:
        print(sample['UserActivityLst'][-5:])

# Check this activity out 
from pprint import pprint

for sample in act_data:
    if sample["AId"] == 2100: # see [8885, 8886, 2100]
        pprint(sample)
        break

# Check if deduplicated 
duplicated_or_not = []
duplicated = []
for sample in usr_data:
    aid = sample["AId"]
    if aid not in duplicated_or_not:
        duplicated_or_not.append(aid)
    else:
        duplicated.append(aid)

duplicated_or_not.sort()
duplicated.sort()

len(duplicated_or_not), len(duplicated)
# There is no deduplicated in this dataset

# Check code status in # The 205 status code tells the client the request was successful and that there is no content to show for it. or 0 
## For user data

usr_code_status = []
for sample in usr_data:
    code = sample["RespCode"]
    if code not in usr_code_status:
        usr_code_status.append(code)

usr_code_status # => 205 is not found

# Check code status
act_code_status = []
for sample in act_data:
    code = sample["RespCode"]
    if code not in act_code_status:
        act_code_status.append(code)

act_code_status # => 205 is not found

# get user code 205 status
usr_code_205_status = []
for sample in usr_data:
    code = sample["RespCode"]
    if code == 205:
        usr_code_205_status.append(sample["AId"])

usr_code_205_status = list(set(sorted(usr_code_205_status))) 
len(usr_code_205_status), usr_code_205_status[-1]

# Check code status
act_code_205_status = []
for sample in act_data:
    code = sample["RespCode"]
    if code == 205:
        act_code_205_status.append(sample["AId"])

act_code_205_status = list(set(sorted(act_code_205_status))) 
len(act_code_205_status), act_code_205_status[-1]
# 4544 - 4141 = 403 is activites deleted

# See which one is deleted
count = 0
for idx in act_code_205_status:
    if idx not in usr_code_205_status:
        count += 1
        print(idx)
        if count == 5:
            break
# Check https://ctsv.hust.edu.vn/#/hoat-dong/8339/chi-tiet

# Check this activity out 
from pprint import pprint

for sample in act_data:
    if sample["AId"] == 8339: 
        pprint(sample)
        break

# Check this activity out 
from pprint import pprint

for sample in usr_data:
    if sample["AId"] == 8339: 
        pprint(sample)
        break

# Only have supervisor for deleted activities?
# See which one is deleted
consider_deleted_activity_id  = []
for idx in act_code_205_status:
    if idx not in usr_code_205_status:
        consider_deleted_activity_id.append(idx)

for sample in usr_data:
    if sample["AId"] in consider_deleted_activity_id: 
        print(sample["AId"], len(sample["UserActivityLst"]))

# Check this activity out 
from pprint import pprint

for sample in act_data:
    if sample["AId"] == 709: 
        pprint(sample)
        break

# Check this activity out 
from pprint import pprint

for sample in usr_data:
    if sample["AId"] == 1405: #8937, 709
        pprint(sample)
        break

for idx in usr_code_205_status:
    if idx not in act_code_205_status:
        print(idx)
# Not inversed

not_exists_event = set(act_code_205_status).union(usr_code_205_status)
len(not_exists_event)

# Basic EDA

# dict: AId first
usr_data_clean = {}
for sample in usr_data:
    if sample["AId"] not in not_exists_event and sample["AId"] not in usr_data_clean:
        usr_data_clean[sample["AId"]] = sample

len(usr_data_clean) # 9085 - 4544

list(usr_data_clean.keys())[-5:]

usr_data_clean[9084]

# dict: AId first
act_data_clean = {}
for sample in act_data:
    if sample["AId"] not in not_exists_event and sample["AId"] not in act_data_clean:
        act_data_clean[sample["AId"]] = sample

len(act_data_clean) # 9085 - 4544

usr_data_clean[9084]

act_data_clean[9084]

from trafilatura.utils import sanitize
from trafilatura import extract

user_info  = {} # UserCode : --> Info, get only relevant fields
activity_info = {} # ACode: --> Info, get only relevant fields

data_combined = []
count = 0

for AId, sample in usr_data_clean.items():
    act_data = act_data_clean[AId]["Activities"][-1] # 8848 duplicated but it's alright now
    activity_info[AId] = {"id": AId,
                          "title": sanitize(act_data["AName"]).strip(), 
                          "type": sanitize(act_data["AType"]).strip(),
                          "desc": extract(act_data["ADesc"]),
                          "btc": act_data["GName"].strip(),
                          "createdTime": act_data["CreateDate"], 
                          "startTime": act_data["StartTime"],
                          "finishTime": act_data["FinishTime"],
                          "deadlineTime": act_data["Deadline"],
                          }
    
    activity_info[AId]["criteria"] = act_data["CriteriaLst"]

    for user_data in sample["UserActivityLst"]:
        user_id = user_data['UserCode'].strip().upper()
        user_info[user_id] = {"id": user_id,
                                            "fullName": user_data["FullName"],
                                            "email": user_data["Email"].strip().lower(),
                                            "faculty": user_data["Faculty"]}
        
        data_combined.append([user_id, str(AId) , str(user_data['UAStatus'])])

len(data_combined), data_combined[0]

len(user_info), user_info['20204876']

len(activity_info), activity_info[8848]

import pandas as pd

df_user_act_interaction = pd.DataFrame(data_combined,
                  columns=["user_id", "activity_id", "accepted_status"])

df_user_act_interaction.head()

# Interaction
with open("output/user_activity_interaction_cleaned.csv", "w") as f:
    df_user_act_interaction.to_csv(f, index=False)

# User Information 
df_user_info = pd.DataFrame.from_dict(list(user_info.values()))
df_user_info.head()


df_user_info.dtypes

with open("output/user_info_cleaned.csv", "w") as f:
    df_user_info.to_csv(f, index=False)

# User Information 
df_act_info = pd.DataFrame.from_dict(list(activity_info.values()))
df_act_info.head()


df_act_info["startTime"] = pd.to_datetime(df_act_info["startTime"])
df_act_info["finishTime"] = pd.to_datetime(df_act_info["finishTime"])
df_act_info["deadlineTime"] = pd.to_datetime(df_act_info["deadlineTime"])

with open("output/activity_info_cleaned.csv", "w") as f:
    df_act_info.to_csv(f, index=False)
