==> notebooks/2_data_deep_cleaning.ipynb <==
# Settings

CLEANED_DATA_OUTPUT = "output/"

# Deep Cleaning

## Read Files

## Input 
import os 
import pandas as pd

with open(os.path.join(CLEANED_DATA_OUTPUT, "user_activity_interaction_cleaned.csv")) as fIn:
    df_interaction = pd.read_csv(fIn)

df_interaction = df_interaction.astype('str')
df_interaction.head()

df_interaction.count()

with open(os.path.join(CLEANED_DATA_OUTPUT, "user_info_cleaned.csv")) as fIn:
    df_user = pd.read_csv(fIn)

df_user.head()

with open(os.path.join(CLEANED_DATA_OUTPUT, "activity_info_cleaned.csv")) as fIn:
    df_activity = pd.read_csv(fIn)

df_activity.head()

df_activity.dtypes

df_interaction_processed = df_interaction.copy()
df_interaction.nunique()

df_user_processed = df_user.copy()
df_user.nunique()
# Email deduplication ??? Later

df_activity_processed = df_activity.copy()
df_activity.nunique()
# The same title ? , Deduplicate desc ?, so many btcs :)

## User Info

df_user_processed.nunique()

df_not_student_user = df_user_processed[df_user_processed["email"].apply(lambda x: not x.endswith("sis.hust.edu.vn"))]
df_not_student_user

df_not_student_user[df_not_student_user["email"].apply(lambda x: not x.endswith("hust.edu.vn"))]

# We removed all not hust students 
df_user_processed = df_user_processed[df_user_processed["email"].apply(lambda x: x.endswith("sis.hust.edu.vn"))]

df_user_processed.isnull().sum()

df_user_processed["faculty"].unique()

df_user_processed[df_user_processed["faculty"].isnull()] # just missing value, it's alright 

# We can fill data for 63188 20211790 20213157
## Search huy.nd151670@sis.hust.edu.vn on team --> Nguyen Dang Huy 20151670 Moi truong 1, check faculty similar user based on https://sami.hust.edu.vn/wp-content/uploads/20162-CK-GT3-TC.pdf
df_user_processed[df_user_processed["id"].isin(["20152600", "20152329"])]

df_user_processed.loc[39790, :] = pd.Series({"id": "20151670", 
                                   "fullName": "Nguyễn Đăng Huy", 
                                   "email": "huy.nd151670@sis.hust.edu.vn", 
                                   "faculty": "Trường Hóa và Khoa học Sự sống"})

df_user_processed.loc[42350, "faculty"] = "Trường Hóa và Khoa học Sự sống"
df_user_processed.loc[51358, "faculty"] = "Viện Kinh tế và Quản lý"


df_user_processed[df_user_processed["faculty"] == "Ban CTSV"]

df_user_processed = df_user_processed[df_user_processed["id"] != "20190000"]

df_user_processed.nunique()

# ? Đepulicated email
df_user_processed[df_user_processed["email"].duplicated(keep=False)].sort_values(by="email")

# Change back to 20198486
temp_id = df_user_processed.loc[35324]["id"]
temp_id

df_interaction_processed[df_interaction_processed["user_id"] == temp_id]

df_user_processed.loc[35324, "id"] = "20198486"
df_interaction_processed.loc[1187660, "user_id"] ="20198486" 

df_user_processed = df_user_processed.drop_duplicates()
df_user_processed[df_user_processed["email"].duplicated(keep=False)].sort_values(by="email")

import re
temp_id_df = df_user_processed["id"].apply(lambda x: re.findall(r'\d+', x)[0]).astype(int)
df_user_processed = df_user_processed[(temp_id_df <= 20240000) & (temp_id_df >= 20120000)]

# some unusual mail ==> All valid 
df_user_processed[df_user_processed["email"].apply(lambda x : "." not in x.split("@")[0] or not bool(re.search(r'\d', x.split("@")[0])))]

df_user_processed

df_user_processed.to_csv("output/user_info_deep_cleaned.csv", index=False)

# Maybe some EDA on number of K65, ... faculty in here

## Activity

df_activity_processed.head()

df_activity_processed.nunique()

df_activity_processed[df_activity_processed["title"].duplicated(keep=False)].sort_values("title")

# 3306, 3293,3285, 3287 deduplicated
df_activity_processed = df_activity_processed[df_activity_processed["id"].apply(lambda x: x not in [3306, 3293, 3285, 3287]) ]
df_activity_processed[df_activity_processed["title"].duplicated(keep=False)].sort_values("title")

df_activity_processed[df_activity_processed["desc"].duplicated(keep=False)].sort_values("desc").dropna() # it's alright

df_activity_processed["id"] = df_activity_processed["id"].astype(str)

df_activity_processed.to_csv("output/activity_info_deep_cleaned.csv", index=False)

## Interaction

df_interaction_processed.dtypes

df_interaction_processed =  df_interaction_processed.merge(df_user_processed, left_on="user_id", right_on="id")[["user_id", "activity_id", "accepted_status"]]
df_interaction_processed

df_interaction_processed = df_interaction_processed.merge(df_activity_processed, left_on= "activity_id", right_on="id")[["user_id", "activity_id", "accepted_status"]]
df_interaction_processed

df_interaction_processed.to_csv("output/user_activity_interaction_deep_cleaned.csv", index=False)
