==> notebooks/7_modeling_hybrid_add_negative.ipynb <==
# Settings

OUTPUT_CLEANED_FOLDER = "output/"

# Read data

import os 
import pandas as pd 

df_interaction = pd.read_csv(os.path.join(OUTPUT_CLEANED_FOLDER, "user_activity_interaction_deep_cleaned.csv"))
df_user = pd.read_csv(os.path.join(OUTPUT_CLEANED_FOLDER, "user_info_deep_cleaned.csv"))
df_activity = pd.read_csv(os.path.join(OUTPUT_CLEANED_FOLDER, "activity_info_deep_cleaned.csv"))

df_activity["id"]  = df_activity["id"].astype(str)

df_activity.head()

df_interaction = df_interaction.astype(str)
df_interaction.dtypes

df_interaction.head()

df_interaction["inter_score"] = 1

df_interaction

df_all = df_interaction.merge(df_activity, left_on=["activity_id"], right_on=["id"])[["user_id", "activity_id", "inter_score", "createdTime"]]
df_all.head()

df_all.tail()

df_all["createdTime"] = pd.to_datetime(df_all["createdTime"])

# Metrics 

import numpy as np
import math

def getHitRatio(ranklist, gtItem):
    if gtItem in ranklist:
        return 1
    return 0

def getNDCG(ranklist, gtItem):
    ar = np.array(ranklist)
    if gtItem in ar:
        return math.log(2) / math.log(np.where(ar == gtItem)[0][0] + 2)
    return 0


# Training Testing Split


def train_test_split_time(df , columns_1 , columns_2 , columns_time):
    '''
    we will use the columns_1 as the first level index
    sort the columns_2 as the target by columns_time
    '''
    
    retrieve_name = df[columns_1].value_counts()[df[columns_1].value_counts() > 5].keys()
    df_test = df[df[columns_1].isin(retrieve_name)]
    df_test = df_test.sort_values(columns_time).groupby(columns_1).tail(1)
    df_train = df.drop(index=df_test.index)
    
    if df_test.shape[0] + df_train.shape[0] == df.shape[0]:
        print('train_test_split succeed!! with df_train shape:(%d,%d), df_test shape:(%d,%d)'
              %(df_train.shape[0],df_train.shape[1],df_test.shape[0],df_test.shape[1])
             )
        return df_train, df_test, retrieve_name
    else:
        print('Oops, something wrong, with df_train shape : (%d,%d), df_test shape : (%d,%d)'
              %(df_train.shape[0],df_train.shape[1],df_test.shape[0],df_test.shape[1])
             )
        print('df_shape : (%d,%d)'
              %(df.shape[0],df.shape[1])
             )

df_train, df_test, retrieve = train_test_split_time(df_all, 'user_id','activity_id', 'createdTime')

df_train[df_train["user_id"] == "20204876"].sort_values(by="createdTime")

df_test[df_test["user_id"] == "20204876"]

We will use the leave-one-out method to evaluate our testing data

Basically, we will generate N samples list that the test user hasn't read and add the ground truth into the list, and see how the recommendation system will score and rank the list

set(df_all["activity_id"].unique())

len(set(df_all["activity_id"].unique()).difference(sorted(df_all.groupby("user_id")['activity_id'].apply(list)["20204876"])))

import random

random.seed(73)

def get_sampling(df_target, df_source, numbers_of_N_sample, generate_negative = False):
    '''
    1.
    return: test_rating in the shape of [ [user_pid,article] * N ]
    2.
    from the target dataframe, sampling the negative sample from source dataframe
    return: list in the shape of [ [((user_id, ground truth)) , [numbers_of_N_sample(negative samples)]] * N]
    '''
    test_rating = df_target[['user_id','activity_id']].values.tolist()
    print('test_rating yield successfully!!!')
    
    if generate_negative:
        act_set = set(df_source["activity_id"].unique())
        mapping_user_act = df_source.groupby("user_id")['activity_id'].apply(list)
        negative_test_rating = []
        for i in range(df_target.shape[0]):
            drop_id = test_rating[i][0]
            list_ = random.sample(list(act_set.difference(mapping_user_act[drop_id])), numbers_of_N_sample)
            negative_test_rating.append([test_rating[i], list_])
        print('negative_test_rating yield successfully!!!')

        return test_rating, negative_test_rating
    else:
        return test_rating

print('''only for generating sample, if there is a sample file already, skip this cell in the future''')
# In the future just execute the next cell to get the test rating & negative sampling
# import time
# import pickle

# time1 = time.time()
# test_rating, negative_test_rating = get_sampling(df_test, df_train, 19, True)
# time2 = time.time()
# print('Took for %d seconds' %(time2-time1))
# with open("output/negative_test_rating_19N.pkl","wb") as f: #in write mode
#     pickle.dump(negative_test_rating, f)

import time
import pickle

time1 = time.time()
test_rating = get_sampling(df_test, df_train, None)
with open("output/negative_test_rating_19N.pkl",'rb') as f: #in read mode, not in write mode, careful
    negative_test_rating = pickle.load(f)
    print('negative_test_rating yield successfully!!!')
    
time2 = time.time()
print('Took for %d seconds' %(time2-time1))

# Create Features for Activity

df_activity_processed = df_activity.copy()
df_activity_processed.nunique()

df_activity_processed.dtypes

df_activity_processed.head()

df_activity_processed["createdTime"]

df_activity_processed["createdTimeYear"] = pd.to_datetime(df_activity_processed["createdTime"]).dt.year

df_activity_processed["createdTimeYear"].unique() # category features

df_activity_processed["type"].unique() # category features

df_activity_processed["btc"].unique() # too much -> concat to title, desc

df_activity_processed["createdTime"].min(), df_activity_processed["createdTime"].max() # categorize based on years

df_activity_processed["criteria"][0]

import re

def clean_json(string):
    p = re.compile('(?<!\\\\)\'')
    return p.sub('\"', string)

import json
citeria_list = {}

for index, row in df_activity_processed.iterrows():
    criteria = row['criteria']
    try:
        json_parsed = json.loads(clean_json(criteria))
        for crite in json_parsed:
                citeria_list[crite["CId"]] = {"id": crite["CId"],
                                            "name": crite["CName"],
                                            "point": crite["CMaxPoint"]}
    except Exception as e:
         pass 
         # Just ignore them, We cannot parse them :)

len(citeria_list) # Categorical 

## Bi-Encoder for encoding text
act_id_to_text = {}

for index, row in df_activity_processed.iterrows():
    title = str(row['title']).strip() if row['title'] is not None else ""
    desc = str(row["desc"]).strip() if row['desc'] is not None else ""
    btc = str(row["btc"]).strip() if row['btc'] is not None else ""
    
    act_id_to_text[row["id"]] = title + "\n\n" + desc + "\n\n" + btc

print("You do not need run this if you already have the embeddings")
# # INPUT TEXT MUST BE ALREADY WORD-SEGMENTED!

# Necessary libraries

# from tqdm import tqdm 
# from sentence_transformers import SentenceTransformer
# from pyvi import ViTokenizer

# SENTENCE_PER_BATCH = 64
# model = SentenceTransformer('bkai-foundation-models/vietnamese-bi-encoder')
# len_all = len(act_id_to_text)

# list_key = []
# list_embedding = []
# count = 0
# batch_processed = []

# for key, value in tqdm(act_id_to_text.items()):
#     list_key.append(key)
#     batch_processed.append(ViTokenizer.tokenize(value))
#     count += 1
#     if not count % SENTENCE_PER_BATCH or count == len_all:
#         embeddings = model.encode(batch_processed)
#         list_embedding.extend(embeddings)
#         batch_processed = []

# act_id_to_emb = {}
# for key, emb in zip(list_key, list_embedding):
#     act_id_to_emb[key] = emb

# import pickle
# with open("output/activity_embedding.pkl", "wb") as fOut:
#     pickle.dump(act_id_to_emb, fOut)

import pickle
with open("output/activity_embedding.pkl", "rb") as fOut:
    act_id_to_emb = pickle.load(fOut)

# Create one-hot vector for type
type_map_one_hot = {}

nunique = df_activity_processed["type"].nunique()
for idx, data in enumerate(df_activity_processed["type"].unique()):
    type_map_one_hot[data] = idx
    vector = [0] * nunique
    vector[idx] = 1.0
    type_map_one_hot[data] = vector

# For startYear
start_year_map_one_hot = {}

nunique = df_activity_processed["createdTimeYear"].nunique()
for idx, data in enumerate(df_activity_processed["createdTimeYear"].unique()):
    start_year_map_one_hot[data] = idx
    vector = [0] * nunique
    vector[idx] = 1.0
    start_year_map_one_hot[data] = vector

# criteria
criteria_map_id = {}
for idx, data in enumerate(list(citeria_list.keys())):
    criteria_map_id[data] = idx

df_activity_processed["criteria"].isnull().sum()

def criteria_map_one_hot(string):
    results = [0] * (len(criteria_map_id) + 1)
    try:
        json_parsed = json.loads(clean_json(string))
        for crite in json_parsed:
            results[criteria_map_id[crite["CId"]]] = 1.0
    except Exception as e:
        results[-1] = 1.0 

    return results

# final features 
act_id_to_features = {} # final features
for _, row in df_activity_processed.iterrows():
    features = []
    activity_id = row["id"]
    # text
    features.extend(act_id_to_emb[activity_id])
    # type one hot 
    features.extend(type_map_one_hot[row["type"]])
    # criteria features
    features.extend(criteria_map_one_hot(row["criteria"]))
    # start year
    features.extend(start_year_map_one_hot[row["createdTimeYear"]])

    act_id_to_features[activity_id] = features

len(act_id_to_features["8300"])

# Create Features for Users

df_user_processed = df_user.copy()
df_user_processed.nunique()

df_user_processed["faculty"].unique()

df_user_processed["faculty"].value_counts(dropna=False)

mapping_similar = {"Trường Cơ khí": "Trường Cơ Khí",
                   "Trường Hóa và Khoa học sự sống": "Trường Hóa và Khoa học Sự sống",
                   "Trường Công nghệ thông tin & Truyền thông": "Trường Công nghệ Thông tin và Truyền thông",
                   "Viện Vật lý kỹ thuật": "Viện Vật lý Kỹ thuật"}


df_user_processed["faculty"] = df_user_processed["faculty"].apply(lambda x :  mapping_similar[x] if x in mapping_similar else x)
df_user_processed["faculty"].value_counts(dropna=False)

# For startYear
faculty_map_one_hot = {}

nunique = df_user_processed["faculty"].nunique(dropna=False)
for idx, data in enumerate(df_user_processed["faculty"].unique()):
    faculty_map_one_hot[data] = idx
    vector = [0] * nunique
    vector[idx] = 1.0
    faculty_map_one_hot[data] = vector

faculty_map_one_hot[np.nan]

df_user_processed

import re
temp_id_df = df_user_processed["id"].apply(lambda x: re.findall(r'\d+', x)[0]).astype(int)

temp_id_df.count(), temp_id_df.drop_duplicates().count() #P vs not P

def school_year_calculate(temp_id):
    bin_list = [20120000, 20130000, 20140000, 20150000,
                20160000, 20170000, 20180000, 20190000,
                20200000, 20210000, 20220000, 20230000,
                20240000]
    
    corresponding_school_year = ["K57", "K58", "K59", "K60",
                                 "K61", "K62", "K63", "K64",
                                  "K65", "K66", "K67", "K68" ]
    
    for i in range(len(bin_list) - 1):
        if temp_id >= bin_list[i] and temp_id < bin_list[i+1]:
            return corresponding_school_year[i] 

df_user_processed["schoolYear"] = temp_id_df.apply(school_year_calculate)

df_user_processed["schoolYear"].value_counts(dropna=False)

# For schoolYear
school_year_map_one_hot = {}

nunique = df_user_processed["schoolYear"].nunique(dropna=False)
for idx, data in enumerate(df_user_processed["schoolYear"].unique()):
    school_year_map_one_hot[data] = idx
    vector = [0] * nunique
    vector[idx] = 1.0
    school_year_map_one_hot[data] = vector

# final features 
user_id_to_features = {} # final features
for _, row in df_user_processed.iterrows():
    features = []
    user_id = row["id"]
    # faculty one hot 
    features.extend(faculty_map_one_hot[row["faculty"]])
    # start year
    features.extend(school_year_map_one_hot[row["schoolYear"]])

    user_id_to_features[user_id] = features

len(user_id_to_features["20204876"])

# Hybrid Recommendation System


user_id_map = {i:v for i, v in enumerate(df_all["user_id"].unique())}
inverse_user_id_map = {v:i for i, v in enumerate(df_all["user_id"].unique())}
activity_id_map = {i:v for i, v in enumerate(df_all["activity_id"].unique())}
inverse_activity_id_map = {v:i for i, v in enumerate(df_all["activity_id"].unique())}

import random 
random.seed(73)
np.random.seed(2023)

df_train_map = df_train.copy()
df_test_map = df_test.copy()
df_all_map = df_all.copy()


df_train_map['user_id'] = df_train_map['user_id'].map(inverse_user_id_map)
df_train_map['activity_id'] = df_train_map['activity_id'].map(inverse_activity_id_map)
df_test_map['user_id'] = df_test_map['user_id'].map(inverse_user_id_map)
df_test_map['activity_id'] = df_test_map['activity_id'].map(inverse_activity_id_map)
df_all_map['user_id'] = df_all['user_id'].map(inverse_user_id_map)
df_all_map['activity_id'] = df_all['activity_id'].map(inverse_activity_id_map)


### creating a mapping list for testing data   
test_rating_map = []
for i in range(len(test_rating)):
    test_rating_map.append([inverse_user_id_map[test_rating[i][0]], inverse_activity_id_map[test_rating[i][1]]])


### creating a mapping list for negative testing data
negative_test_rating_map = []
for i in range(len(negative_test_rating)):
    negative_test_rating_map.append([
                                     [inverse_user_id_map[negative_test_rating[i][0][0]],
                                      inverse_activity_id_map[negative_test_rating[i][0][1]]],
                                     list(map(lambda w: inverse_activity_id_map[w], negative_test_rating[i][1]))])

### append negative sample into df_training_map

NEGATIVE_FACTOR = 5 # 50 

negative_training_size = NEGATIVE_FACTOR * df_train_map.shape[0]

ng_user_id = np.random.randint(low = 0, 
                               high = max(user_id_map.keys()), 
                               size = negative_training_size)

ng_act_id = np.random.randint(low = 0, 
                              high = max(activity_id_map.keys()), 
                              size = negative_training_size) # including false negative ??

act_set = set(df_all["activity_id"].unique())
mapping_user_act = df_all.groupby("user_id")['activity_id'].apply(list)

df_ngt = pd.DataFrame(columns=['user_id','activity_id', 'inter_score'],
                      data=np.array([ng_user_id, ng_act_id, [0] * negative_training_size]).T)

df_train_map = df_train_map.drop_duplicates(subset=['user_id','activity_id'], keep = 'first')
df_train_map = pd.concat([df_train_map[['user_id','activity_id', 'inter_score']], df_ngt] , ignore_index=True).reset_index()[["user_id", 'activity_id', 'inter_score']]

# df_train_map = df_train_map.sort_values(by="inter_score", ascending=False)
# df_train_map

# df_train_map = df_train_map.drop_duplicates(["user_id", "activity_id"], keep= "first") # Not quite right, negative samples could be dedup but maybe it alright
# df_train_map

# Modeling

# https://github.com/TeddyWang0202/recommendation_system/blob/master/Hybrid.ipynb

import tensorflow as tf
gpu_num = 7
gpus = tf.config.list_physical_devices('GPU')
tf.config.set_visible_devices(gpus[gpu_num], 'GPU')

import keras
keras.__version__
# install tensorflow[and-cuda] to enable keras 2.15.0

from keras.layers import Embedding

num_items = len(act_id_to_emb.items())

### Try create embedding matrix
activity_vector_create = act_id_to_emb.copy()
activity_vector_create[-1] = [0.0] * 768

activity_map_vector = {(inverse_activity_id_map[k] if k in inverse_activity_id_map else -1): v  for k, v in activity_vector_create.items()}
activity_map_vector.pop(-1)

print(len(activity_map_vector))
input_len = 768
embedding_matrix = np.zeros((num_items, input_len), dtype=np.float32)
for i in activity_map_vector.keys():
    embedding_matrix[i] = np.array(activity_map_vector[i][:input_len])

bert_embedding_layer = Embedding(num_items,
                            input_len,
                            weights=[embedding_matrix],
                            name = 'BERT_act_embedding_vector',
                            trainable=False)


from keras import initializers
from keras.models import Model
from keras.regularizers import l2
from keras.layers import multiply, Input, Reshape, Flatten, Dense, Concatenate


def init_normal(shape, dtype=None):
    return initializers.VarianceScaling(scale=0.01, 
    mode='fan_in', distribution='normal')(shape)

def get_model(num_users, num_items, latent_dim, embedding_layer, regs=[0, 0]):
    
    ''' Deep part : user-item-log embedding part '''
    ### define placeholder.
    user_id_input = Input(shape=[1], name='user')
    item_id_input = Input(shape=[1], name='item')

    ### define embedding size and layers.

    user_embedding = Embedding(output_dim = latent_dim, 
                               input_dim = num_users,
                               input_length= 1, 
                               name='user_embedding',
                               embeddings_regularizer = l2(regs[0]),
                               embeddings_initializer = init_normal)(user_id_input)
    
    item_embedding = Embedding(output_dim = latent_dim, 
                               input_dim = num_items,
                               input_length=1, 
                               name='item_embedding',
                               embeddings_regularizer = l2(regs[1]),
                               embeddings_initializer = init_normal)(item_id_input)

    user_vecs = Reshape([latent_dim])(user_embedding)
    item_vecs = Reshape([latent_dim])(item_embedding)
    
    mf = multiply([user_vecs, item_vecs])

    ''' Wide part : item embedding vector part '''
    ### input extracted matrix(here use embedding) and concatenate with one single DNN layer 
    item_vector = Flatten()(embedding_layer(item_id_input))
    item_vector_dense = Dense(384, 
                              activation='relu')(item_vector)
    
    '''Concatenate the deep and wide part'''
    
    concate_vector = Concatenate()([mf,
                                    item_vector_dense,
                                   ])
    
    ### The prediction, which we calculate the loss function with ground truth and optimize.
    y_hat= Dense(1, 
                 activation='selu', 
                 kernel_initializer='lecun_uniform', 
                 name = 'prediction')(concate_vector)
    
    
    model = Model(inputs=[user_id_input, item_id_input], 
                  outputs=y_hat)
    
    return model


num_users = len(user_id_map.items())
num_items = len(activity_id_map.items())

topK = 5
latent_dim = 15
epochs = 100
batch_size = 4096
evaluation_threads = 1
best_hr, best_ndcg, best_iter = -1, -1, -1
model_out_file = 'models/DWN_%d_%d.h5' %(latent_dim, time.time())


learning_rate = 1e-4
model = get_model(num_users, 
                  num_items, 
                  latent_dim, 
                  bert_embedding_layer, 
                  regs=[0, 0])


from keras.optimizers import Adam
import heapq


model.compile(optimizer=Adam(learning_rate=learning_rate), loss='bce')
hr_list = []
loss_list = []
patience = 5
early_stop =True

# Generate training instances
user_input, item_input = df_train_map["user_id"], df_train_map["activity_id"]
labels = df_train_map["inter_score"]
# create user_test_input, 
user_test_input = [] 
item_test_input = []

for data_sample in negative_test_rating_map:
    user_id = data_sample[0][0]
    items_id = [data_sample[0][1]] +  data_sample[1]
    user_test_input.extend([user_id] * len(items_id))
    item_test_input.extend(items_id)

user_test_input = np.array(user_test_input)
item_test_input = np.array(item_test_input)


print('start training...')
for epoch in range(epochs):
    t1 = time.time()
    print('training for the %d epoch' % (epoch))    
    # Training
    hist = model.fit([np.array(user_input), np.array(item_input)], #input
                         np.array(labels), # labels 
                         batch_size = batch_size,
                         validation_data = ([df_test_map["user_id"],df_test_map["activity_id"]],
                                            df_test_map["inter_score"]) ,
                         epochs = 1, 
                         shuffle = True)
    
    t2 = time.time()

    predictions = model.predict([user_test_input,
                            item_test_input], 
                            batch_size=batch_size)

    # Evaluation
    hits, ndcgs = [], []

    df = pd.DataFrame(np.concatenate([user_test_input.reshape(-1, 1), predictions], axis=1), 
             columns=["user_id", "prediction"])

    df["user_id"] = df["user_id"].astype(int)

    user_predictions = {} 
    for index, row in df.iterrows():
        user_predictions.setdefault(int(row["user_id"]), []).append(row["prediction"])    

    for idx in range(len(test_rating_map)):
        rating = test_rating_map[idx]
        ### rating should be like (  [user_id,article_pid(only one article, which is ground truth)]  ) 
        items = negative_test_rating_map[idx][1]
        ### items should be like (  [article_pid(199 ones, which are ones the reader haven't read)]  )
        u = rating[0]
        gtItem = rating[1]
        items.append(gtItem)
        map_item_score = {}
        for i in range(len(items)):
            item = items[i]
            map_item_score[item] = user_predictions[u][i]
        
        items.pop() # LOL 
        ranklist = heapq.nlargest(topK, map_item_score, key=map_item_score.get)

        hr = getHitRatio(ranklist, gtItem)
        ndcg = getNDCG(ranklist, gtItem)

        hits.append(hr)
        ndcgs.append(ndcg)

    hr, ndcg, loss = np.array(hits).mean(), np.array(ndcgs).mean(), hist.history['val_loss'][0]    
    hr_list.append(hr)
    loss_list.append(loss)
    # Using patience to set the early stopping.
    # Always to save the model with minimun loss.
    if hr < np.max(hr_list):
        patience_count += 1
    else:
        patience_count = 0
        best_hr, best_ndcg, best_iter = hr, ndcg, epoch
        model.save_weights(model_out_file, 
                           overwrite=True)
        
    if (early_stop) and (patience_count == patience):
        break
    if epoch % 1 == 0:
        print('Iteration %d [%.1f s]: HR = %.4f, NDCG = %.4f, val_loss = %.4f [%.1f s]' 
            % (epoch,  t2-t1, hr, ndcg, loss, time.time()-t2))

print("End. Best Iteration %d:  HR = %.4f, NDCG = %.4f. " %(best_iter, best_hr, best_ndcg))
